+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Make theme, key bindings and options configurable through a configuration file.
+
 ### Changed
 - `o` key binding opens the channel page in browser if the channels is the selected block.